Made program a bit 'fairer' - quests generate a 'cancelled' key if ended early when not locked in fixed a bug with one time schedules not deactivating after finishing Changed many TIAB related values to floats for QoL. Now 1 minute of duration can cause an equal quest reward difference regardless of whether its 53 to 54 minutes or crossing the 1 hour threshold.

+++ PQiOS/PQPrototype/Persistence/Entities/TimeInABottle+CoreDataProperties.swift
@@ -19,8 +19,8 @@ extension TimeInABottle {
     @NSManaged public var weeklyTimeLimit: Int16
     @NSManaged public var weeklyTimeReset: Date?
     @NSManaged public var weeklyLimitIncreaseDate: Date?
-    @NSManaged public var weeklyTimeCollected: Int16
-    @NSManaged public var timeStored: Int64
+    @NSManaged public var weeklyTimeCollected: Float
+    @NSManaged public var timeStored: Float
 
 }
